requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
httpx>=0.23.0
python-dotenv>=0.19.0
pydantic>=1.10.0
pydantic-settings>=2.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
redis>=5.0.0
python-dateutil>=2.8.2
playwright>=1.44.0
elevenlabs>=1.0.0
beautifulsoup4>=4.12.0
